refactor: move ui assets to ui package
