+++ apps/frontend/src/app/layout.tsx
@@ -2,7 +2,6 @@ import { UIProvider } from "@repo/ui/providers"
 import { ColorModeScript } from "@yamada-ui/react"
 import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
-import { Footer } from "@/components/Generic/Footer"
 
 const geistSans = Geist({
   subsets: ["latin"],
