+++ packages/ui/src/components/Generic/Footer/FooterDecoration.tsx
@@ -1,7 +1,7 @@
-import { Image } from "@repo/ui/components/Primitive"
 import { Motion } from "@yamada-ui/react"
 import { memo } from "react"
-import Shuttle from "@/assets/shuttle.png"
+import Shuttle from "@/components/Asset/shuttle.png"
+import { Image } from "@/components/Primitive"
 
 export const FooterDecoration = memo(() => {
   return (
